requirements.txt
flask
tensorflow-cpu
gunicorn
matplotlib
opencv-contrib-python-headless